mcp/requirements.txt
mcp[cli]>=1.9.0
azure-search-documents>=11.6.0
azure-identity>=1.19.0
python-dotenv>=1.1.0
uvicorn>=0.34.0
